% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-jags.R
\name{fetch_jags}
\alias{fetch_jags}
\title{Fetch JAGS files}
\usage{
fetch_jags(
  dir,
  vars = NULL,
  elevation_span = NULL,
  exclusion_zone = NULL,
  std_from = c("top", "bottom", "none"),
  land_type = FALSE,
  ...
)
}
\arguments{
\item{dir}{Directory fetch files from.}

\item{vars}{Explanatory variables to use.}

\item{elevation_span}{Elevation spans to use.}

\item{exclusion_zone}{Exclusion zones to use.}

\item{std_from}{Side of standardization to use. One of \code{top}, \code{bottom} or
\code{none}.}

\item{land_type}{If \code{TRUE}, select files from analyses performed by land type.}

\item{...}{Other arguments passed to \code{\link[=list_files]{list_files()}}.}
}
\value{
A vector of file dirs with class \code{collection}.
}
\description{
Fetch JAGS files produced by \code{\link[=run_jags]{run_jags()}}.
}
