% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-auto.R
\name{read_auto}
\alias{read_auto}
\title{Read multiple delimited files}
\usage{
read_auto(files, rename = NULL, ..., names_to = rlang::zap())
}
\arguments{
\item{files}{A list of files, connections, or literal data.}

\item{rename}{A named vector of regex and replacements used to rename columns
in each file.}

\item{...}{One or more arguments separated with a comma to pass on to
\code{\link[data.table:fread]{data.table::fread()}}.}

\item{names_to}{Either a string or NULL. See \code{\link[purrr:map_dfr]{purrr::map_df()}} for details.}

\item{merge}{Should the data be merged in a single data frame?}

\item{clean_names}{Should the column names be cleaned?}

\item{ignore_case}{Should letter case be ignored?}
}
\description{
Wrapper around \code{\link[data.table:fread]{data.table::fread()}} to read multiple files.
}
