% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trs-package.R
\docType{package}
\name{trs-package}
\alias{trs}
\alias{trs-package}
\title{trs: The Temperature Range Squeeze Hypothesis}
\description{
Code to reproduce the results behind the paper "Diurnal temperature range as a key predictor of plants' elevation ranges globally".
}
\keyword{internal}
